fix: browser bundling
In botframework-streaming, we used browserify to
produce a browser friendly bundle. Unfortunately,
that makes it difficult to ensure we don't pull in
transitive dependencies within the resulting
bundle. Using esbuild and an explicit set of
browser package rewrites is a little better. This
works for adaptive-expressions, too.

This also removes browser bundling for
botframework-connector, which was never really
useful anyways. Though @azure/ms-rest-js claims to
be isomorphic, it is not trivially compatible with
esbuild bundling.